+++ src/main/java/com/rdp/printing/finish/entity/UserFinish.java
@@ -7,8 +7,6 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -20,14 +18,14 @@ public class UserFinish extends BaseEntity {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long userFinishId;
     private String cutting;
-    private BigDecimal cuttingAmount;
+    private Long cuttingAmount;
     private String digitalDyeCutting;
     private Integer dyeCuttingSheets;
-    private BigDecimal dyeCuttingAmount;
+    private Long dyeCuttingAmount;
     private String laminationType;
     private Integer laminationSheets;
-    private BigDecimal laminationAmount;
+    private Long laminationAmount;
     private Integer totalSheets;
-    private BigDecimal totalAmount;
+    private Long totalAmount;
     private String side;
 }
